refactor: mode debug tools into webgpu debugger
